Test(tutorial5) : Tried another type declaration

+++ tutorial5/src/main.rs
@@ -1,6 +1,6 @@
 fn main() {
-    let x = 25_i8;
-    let y = 10_i8;
+    let x = 25 as i8;
+    let y = 10 as i8;
 
     let z = x / y;
     println!("{}", z); // Finally gives 25.5
